refactor: 경로 정리, 불필요한 코드 정리
