+++ src/constants/delimiters.js
@@ -1,8 +1,2 @@
+// eslint-disable-next-line
 export const NUMBER_DELIMITER = ',';
-
-export const INPUT_DELIMITER = '> ';
-
-export const SQUR_BRACKETS = Object.freeze({
-  left: '[',
-  right: ']',
-});
