+++ src/constants/index.js
@@ -1,3 +1,3 @@
-export * from './rules';
-export * from './messages';
 export * from './delimiters';
+export * from './messages';
+export * from './rules';
