+++ src/constants/messages.js
@@ -1,11 +1,13 @@
-import { INPUT_DELIMITER, NUMBER_DELIMITER } from './delimiters';
+import { NUMBER_DELIMITER } from './delimiters';
 import { LOTTO_RULE, RESTART_KEY } from './rules';
 
+const INPUT_QUERY_PREFIX = '> ';
+
 export const INPUT_MESSAGES = Object.freeze({
-  paymentAmount: `${INPUT_DELIMITER}구입금액을 입력해 주세요.`,
-  winningLottoNumbers: `\n${INPUT_DELIMITER}당첨 번호를 입력해 주세요.`,
-  bonusNumber: `\n${INPUT_DELIMITER}보너스 번호를 입력해 주세요.`,
-  restart: `\n${INPUT_DELIMITER}다시 시작하시겠습니까? (y/n)`,
+  paymentAmount: `${INPUT_QUERY_PREFIX}구입금액을 입력해 주세요.`,
+  winningLottoNumbers: `\n${INPUT_QUERY_PREFIX}당첨 번호를 입력해 주세요.`,
+  bonusNumber: `\n${INPUT_QUERY_PREFIX}보너스 번호를 입력해 주세요.`,
+  restart: `\n${INPUT_QUERY_PREFIX}다시 시작하시겠습니까? (y/n)`,
 });
 
 export const OUTPUT_MESSAGES = Object.freeze({
@@ -15,11 +17,11 @@ export const OUTPUT_MESSAGES = Object.freeze({
   countUnit: '개',
   profitUnit: '%',
 });
-('');
+
 const ERROR_PREFIX = '[ERROR]';
 
 export const ERROR_MESSAGE = Object.freeze({
-  inValidwinningNumbersForm: `${ERROR_PREFIX} 당첨 번호 숫자를 쉼표("${NUMBER_DELIMITER}")로 구분하여 입력해 주세요.\n`,
+  inValidWInningNumbersForm: `${ERROR_PREFIX} 당첨 번호 숫자를 쉼표("${NUMBER_DELIMITER}")로 구분하여 입력해 주세요.\n`,
   alreadyInLottoNumber: `${ERROR_PREFIX} 보너스 번호는 당첨 번호에 없는 번호여야 합니다.\n`,
   notInteger: `${ERROR_PREFIX} 정수가 아닙니다.\n`,
   inDivisibleByPrice: `${ERROR_PREFIX} 구매 금액이 ${LOTTO_RULE.price.toLocaleString('ko-KR')}${OUTPUT_MESSAGES.moneyUnit} 단위가 아닙니다.\n`,
