+++ src/domains/Statistics.js
@@ -11,11 +11,7 @@ class Statistics {
   get profitRate() {
     return this.#reward.profitRate;
   }
-
-  // get totalPrizes() {
-  //   return this.#reward.totalPrizes;
-  // }
-
+  
   get statisticsResult() {
     return this.#ranks.reduce(
       (acc, rank) => {
