+++ src/domains/index.js
@@ -1,7 +1,7 @@
-import Validator from './Validator';
 import LottoGame from './LottoGame';
 import LottoMachine from './LottoMachine';
 import Statistics from './Statistics';
+import Validator from './Validator';
 import WinningLotto from './WinningLotto';
 
 export { Validator, LottoMachine, LottoGame, Statistics, WinningLotto };
