+++ src/services/GameController.js
@@ -1,8 +1,6 @@
-import { RESTART_KEY, OUTPUT_MESSAGES, WINNING_RULE } from '../constants';
+import { RESTART_KEY } from '../constants';
 import { LottoGame, Validator } from '../domains';
-import { Console } from '../utils';
-import InputView from '../views/InputView';
-import OutputView from '../views/OutputView';
+import { InputView, OutputView } from '../views';
 import InputController from './InputController';
 
 class GameController {
@@ -46,18 +44,16 @@ class GameController {
 
   async #getValidWinningLottoNumbers() {
     return await InputController.retryOnInvalidInput(async () => {
-      //입력값 가져오기
       const lottoNumbersInput = await InputView.readWinningLottoNumbers();
-      // 유효성 검사
+
       this.#lottoGame.winningLottoNumbers = lottoNumbersInput;
     });
   }
 
   async #getValidBonusNumber() {
     return await InputController.retryOnInvalidInput(async () => {
-      //입력값 가져오기
       const bonusNumberInput = await InputView.readBonusNumber();
-      // 유효성 검사
+
       this.#lottoGame.bonusNumber = bonusNumberInput;
     });
   }
@@ -70,7 +66,7 @@ class GameController {
 
   async #restartLottoGameAction() {
     const restartInput = await InputView.readRestart();
-    Validator.chaeckRestartForm(restartInput);
+    Validator.checkRestartForm(restartInput);
 
     if (restartInput === RESTART_KEY.restart) {
       this.#lottoGame = new LottoGame();
