+++ src/services/index.js
@@ -1,0 +1,5 @@
+import GameController from "./GameController";
+import InputController from "./InputController";
+
+
+export {GameController, InputController}
