+++ src/views/index.js
@@ -1,0 +1,4 @@
+import OutputView from './OutputView';
+import InputView from './InputView';
+
+export { OutputView, InputView };
